move AddProjectUser.svelte into the project view folder. Apply feedback from PR.
