+++ frontend/src/lib/forms/index.ts
@@ -7,11 +7,6 @@ import { superForm, type EnhancedForm } from 'sveltekit-superforms/client';
 import type { AnyZodObject, z } from 'zod';
 import {superValidate} from "sveltekit-superforms/server";
 
-function validate_email(email: string) {
-	// https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/email#basic_validation
-	return /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/.test(email)
-}
-
 //we've got to wrap this in our own version because we're not using the server side component, which this expects
 export function lexSuperForm<S extends AnyZodObject>(schema: S): EnhancedForm<S, string> {
 	return superForm(undefined, {validators: schema});
@@ -29,6 +24,5 @@ export {
 	ProtectedForm,
 	type Token,
 	Input,
-	validate_email,
 	Select
 }
