+++ frontend/src/routes/project/[project_code]/+page.ts
@@ -1,14 +1,14 @@
 import { getClient } from "$lib/graphQLClient";
 import type { PageLoadEvent } from "./$types";
 import { graphql } from "$lib/gql";
-import type { ProjectPageQuery } from "$lib/gql/graphql";
+import type { AddProjectMemberInput, ProjectPageQuery } from "$lib/gql/graphql";
 import { invalidate } from "$app/navigation";
 
 export type ProjectUser = ProjectPageQuery["projects"][0]["ProjectUsers"][number];
 
 export async function load(event: PageLoadEvent) {
     const client = getClient(event);
-    const project_code = event.params.project_code;
+    const projectCode = event.params.project_code;
     const result = await client.query(graphql(`
         query projectPage($projectCode: String!) {
             projects(where: {code: {_eq: $projectCode}}) {
@@ -30,13 +30,39 @@ export async function load(event: PageLoadEvent) {
                 }
             }
         }
-`), { projectCode: project_code }).toPromise();
+`), { projectCode }).toPromise();
     if (result.error) throw new Error(result.error.message);
     event.depends(`project:${result.data?.projects[0]?.id}`);
     return {
         project: result.data?.projects[0],
-        code: project_code
+        code: projectCode
     };
+}
+
+export async function _addProjectUser(input: AddProjectMemberInput) {
+    //language=GraphQL
+    const result = await getClient().mutation(
+        graphql(`
+            mutation AddProjectUser($input: AddProjectMemberInput!) {
+                addProjectMember(input: $input) {
+                    project {
+                        id
+                    }
+                    errors {
+                        ... on Error {
+                            message
+                        }
+                    }
+                }
+            }
+        `),
+        {input: input}, 
+        //invalidates the graphql project cache
+        {additionalTypenames: ['Projects']}
+    ).toPromise();
+    if (!result.error)
+        invalidate(`project:${input.projectId}`);
+    return result;
 }
 
 export async function _deleteProjectUser(projectId: string, userId: string) {
